backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
openai==1.50.0
pypdf2==3.0.1
spacy==3.7.6
pydantic==2.11.0
python-dotenv==1.0.1
httpx==0.27.2  # Updated to resolve conflict